szu-rag-crawler/requirements.txt
scrapy>=2.11
httpx>=0.27
beautifulsoup4>=4.12
lxml>=5.1
markdownify>=0.12
fastapi>=0.110
uvicorn>=0.29
python-multipart>=0.0.9
apscheduler>=3.10,<4.0
pydantic-settings>=2.2
pyyaml>=6.0
python-dotenv>=1.0
parsel>=1.8
